+++ admittance_controller/include/admittance_controller/admittance_rule_impl.hpp
@@ -23,10 +23,12 @@
 #include <string>
 #include <vector>
 
-#include <control_toolbox/filters.hpp>
-#include <tf2_eigen/tf2_eigen.hpp>
-
+#include "geometry_msgs/msg/transform_stamped.hpp"
 #include "rclcpp/duration.hpp"
+#include "rclcpp/utilities.hpp"
+#include "tf2_eigen/tf2_eigen.hpp"
+#include "tf2_geometry_msgs/tf2_geometry_msgs.hpp"
+#include "tf2_ros/transform_listener.h"
 
 namespace admittance_controller
 {
@@ -107,8 +109,7 @@ controller_interface::return_type AdmittanceRule::reset(const size_t num_joints)
   admittance_state_ = AdmittanceState(num_joints);
 
   // reset forces
-  wrench_world_.setZero();
-  end_effector_weight_.setZero();
+  wrench_base_.setZero();
 
   // load/initialize Eigen types from parameters
   apply_parameters_update();
@@ -121,8 +122,6 @@ void AdmittanceRule::apply_parameters_update()
   parameter_handler_->try_get_params(parameters_);
 
   // update param values
-  end_effector_weight_[2] = -parameters_.gravity_compensation.CoG.force;
-  vec_to_eigen(parameters_.gravity_compensation.CoG.pos, cog_pos_);
   vec_to_eigen(parameters_.admittance.mass, admittance_state_.mass);
   vec_to_eigen(parameters_.admittance.stiffness, admittance_state_.stiffness);
   vec_to_eigen(parameters_.admittance.selected_axes, admittance_state_.selected_axes);
@@ -162,35 +161,34 @@ controller_interface::return_type AdmittanceRule::update(
     reference_joint_state.positions, parameters_.ft_sensor.frame.id, tf);
   admittance_state_.ref_trans_base_ft = tf;
 
-  // --- world frame to base frame (we only need the rotation) ---
-  success &= kinematics_->calculate_link_transform(
-    current_joint_state.positions, parameters_.fixed_world_frame.frame.id, tf);
-  const Eigen::Matrix3d rot_world_base = tf.rotation();
-
   // --- control/base frame to base frame (rotation only) ---
   success &= kinematics_->calculate_link_transform(
     current_joint_state.positions, parameters_.control.frame.id, tf);
   admittance_state_.rot_base_control = tf.rotation();
 
   success &= kinematics_->calculate_link_transform(
-    current_joint_state.positions, parameters_.gravity_compensation.frame.id, tf);
-  const Eigen::Matrix3d rot_tf_cog = tf.rotation();
-
-  success &= kinematics_->calculate_link_transform(
     current_joint_state.positions, parameters_.ft_sensor.frame.id, tf);
   const Eigen::Matrix3d rot_tf_base_ft = tf.rotation();
 
-  // wrench processing (gravity + filter) in world
-  process_wrench_measurements(
-    measured_wrench,
-    // pass rotations into sensor and CoG:
-    rot_world_base * rot_tf_base_ft, rot_world_base * rot_tf_cog);
+  // process the wrench measurements, expect the result in ft_sensor.frame (=FK-accessible frame)
+  if (!process_wrench_measurements(measured_wrench))
+  {
+    desired_joint_state = reference_joint_state;
+    return controller_interface::return_type::ERROR;
+  }
+  // fill the Wrench (there is no fromMsg for wrench)
+  wrench_filtered_ft_(0) = measured_wrench_filtered_.wrench.force.x;
+  wrench_filtered_ft_(1) = measured_wrench_filtered_.wrench.force.y;
+  wrench_filtered_ft_(2) = measured_wrench_filtered_.wrench.force.z;
+  wrench_filtered_ft_(3) = measured_wrench_filtered_.wrench.torque.x;
+  wrench_filtered_ft_(4) = measured_wrench_filtered_.wrench.torque.y;
+  wrench_filtered_ft_(5) = measured_wrench_filtered_.wrench.torque.z;
 
-  // transform filtered wrench into the robot base frame
+  // Rotate (and not transform) to the base frame, explanation lower
   admittance_state_.wrench_base.block<3, 1>(0, 0) =
-    rot_world_base.transpose() * wrench_world_.block<3, 1>(0, 0);
+    rot_tf_base_ft * wrench_filtered_ft_.block<3, 1>(0, 0);
   admittance_state_.wrench_base.block<3, 1>(3, 0) =
-    rot_world_base.transpose() * wrench_world_.block<3, 1>(3, 0);
+    rot_tf_base_ft * wrench_filtered_ft_.block<3, 1>(3, 0);
 
   // populate current joint positions in the state
   vec_to_eigen(current_joint_state.positions, admittance_state_.current_joint_pos);
@@ -230,8 +228,8 @@ bool AdmittanceRule::calculate_admittance_rule(AdmittanceState & admittance_stat
   Eigen::Matrix<double, 3, 3> K_rot = Eigen::Matrix<double, 3, 3>::Zero();
   K_pos.diagonal() = admittance_state.stiffness.block<3, 1>(0, 0);
   K_rot.diagonal() = admittance_state.stiffness.block<3, 1>(3, 0);
-  // Transform to the control frame
-  // A reference is here:  https://users.wpi.edu/~jfu2/rbe502/files/force_control.pdf
+  // Rotate (and not transform) to the control frame, explanation can be found
+  // in the reference here:  https://users.wpi.edu/~jfu2/rbe502/files/force_control.pdf
   // Force Control by Luigi Villani and Joris De Schutter
   // Page 200
   K_pos = rot_base_control * K_pos * rot_base_control.transpose();
@@ -271,9 +269,14 @@ bool AdmittanceRule::calculate_admittance_rule(AdmittanceState & admittance_stat
 
   // zero out any forces in the control frame
   Eigen::Matrix<double, 6, 1> F_control;
+  // rotate to control frame
   F_control.block<3, 1>(0, 0) = rot_base_control.transpose() * F_base.block<3, 1>(0, 0);
   F_control.block<3, 1>(3, 0) = rot_base_control.transpose() * F_base.block<3, 1>(3, 0);
+
+  // zero out forces/torques for axes not in selected_axes
   F_control = F_control.cwiseProduct(admittance_state.selected_axes);
+
+  // rotate to base frame
   F_base.block<3, 1>(0, 0) = rot_base_control * F_control.block<3, 1>(0, 0);
   F_base.block<3, 1>(3, 0) = rot_base_control * F_control.block<3, 1>(3, 0);
 
@@ -306,32 +309,29 @@ bool AdmittanceRule::calculate_admittance_rule(AdmittanceState & admittance_stat
   return success;
 }
 
-void AdmittanceRule::process_wrench_measurements(
-  const geometry_msgs::msg::Wrench & measured_wrench,
-  const Eigen::Matrix<double, 3, 3> & sensor_world_rot,
-  const Eigen::Matrix<double, 3, 3> & cog_world_rot)
+bool AdmittanceRule::process_wrench_measurements(const geometry_msgs::msg::Wrench & measured_wrench)
 {
-  Eigen::Matrix<double, 3, 2, Eigen::ColMajor> new_wrench;
-  new_wrench(0, 0) = measured_wrench.force.x;
-  new_wrench(1, 0) = measured_wrench.force.y;
-  new_wrench(2, 0) = measured_wrench.force.z;
-  new_wrench(0, 1) = measured_wrench.torque.x;
-  new_wrench(1, 1) = measured_wrench.torque.y;
-  new_wrench(2, 1) = measured_wrench.torque.z;
-
-  // transform to world frame
-  Eigen::Matrix<double, 3, 2> new_wrench_base = sensor_world_rot * new_wrench;
-
-  // apply gravity compensation
-  new_wrench_base(2, 0) -= end_effector_weight_[2];
-  new_wrench_base.block<3, 1>(0, 1) -= (cog_world_rot * cog_pos_).cross(end_effector_weight_);
-
-  // apply smoothing filter
-  for (Eigen::Index i = 0; i < 6; ++i)
+  // pass the measured_wrench in its ft_sensor measurement_frame
+  measured_wrench_.header.frame_id = parameters_.ft_sensor.measurement_frame.id;
+  measured_wrench_.wrench = measured_wrench;
+  // output should be ft_sensor frame (because kinematics is only up to there)
+  measured_wrench_filtered_.header.frame_id = parameters_.ft_sensor.frame.id;
+  // apply the filter
+  auto ret = filter_chain_->update(measured_wrench_, measured_wrench_filtered_);
+  // check the output wrench was correctly transformed into the desired frame
+  if (measured_wrench_filtered_.header.frame_id != parameters_.ft_sensor.frame.id)
   {
-    wrench_world_(i) = filters::exponentialSmoothing(
-      new_wrench_base(i), wrench_world_(i), parameters_.ft_sensor.filter_coefficient);
+    RCLCPP_ERROR_ONCE(
+      rclcpp::get_logger("AdmittanceRule"),
+      "Wrench frame transformation missing.\n"
+      "If ft_sensor.frame != ft_sensor.measurement_frame, kinematics has no idea about "
+      "ft_sensor.measurement_frame.\n"
+      "Fix the frames or provide a filter that transforms the wrench from "
+      "ft_sensor.measurement_frame"
+      " to ft_sensor.frame");
+    return false;
   }
+  return ret;
 }
 
 const control_msgs::msg::AdmittanceControllerState & AdmittanceRule::get_controller_state()